Rewrite to use array
Rewritten the whole script so that initial values are entered
into an array. This was initially done to solve the issue of
adding the 'dot' while making sure it is added only once and
using it correctly on initial setup (when 0 is displayed and
by pressing '.' we want '0.' to be displayed). However, it
helped to clear and tidy up the whole script.

Added a few comments to aid visualisation of what is to be expected.