Create functions for number, fix ops and =
Created the first set of functions for the number operations to
tidy up repetitive code.
Noticed bug in consequent operations. To fix: * removed most code
from operations if totalled (all is needed is to remember the
operator pressed and wait for further events); * added extra
condition for equal so that once a calculation has been totalled
and only an operator button has been pressed, it then will operate
on itself (i.e. when '+' pressed it will double)