Fix too large numbers temporarily
Added a temporary solution for numbers to large to display so when e.g.
multiplying 999999999*999999999=*999999999 we do not get 0 as a result.

It now shows 'too large' as a result, but it still works for further
calculations of that number, and will correctly display as long as
result is smaller than 10 digits.